% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillkernel.R
\name{chillkernel}
\alias{chillkernel}
\title{Perform a two-dimensional kernel density estimation for yield and chill}
\usage{
chillkernel(chill, yield)
}
\arguments{
\item{chill}{is a list of observed annual chill portions corresponding to another list with annual yields.}

\item{yield}{is a list of observed annual yields corresponding to another list with annual chill portions.}
}
\description{
The function produces a matrix of the estimated density (z) of yield (y) and chill (x). 
As the density function restricts the shape of the kernel to a bivariate normal kernel, it looks slightly different compared to the scatter plot estimates in the pasitR::chillscatter() function.
Density surface plot of chill portions (x) and yield (y). The legend shows the value for the estimated density (z).
}
\examples{
chill <- sample(x = 1:50, size = 20, replace = TRUE)
yield <- sample(x = 1000:5000, size = 20, replace = TRUE)
chillkernel(chill, yield)

}
\keyword{chill}
\keyword{density}
\keyword{kernel}
\keyword{yield}
