% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillkernelslicerange.R
\name{chillkernelslicerange}
\alias{chillkernelslicerange}
\title{Estimated yield given the expected chill, based on a slice of 'z' from the Kernel density plot of chill and yield data.}
\usage{
chillkernelslicerange(chill, yield, min_chill, max_chill)
}
\arguments{
\item{chill}{is a list of observed annual chill portions corresponding to another list with annual yields.}

\item{yield}{is a list of observed annual yields corresponding to another list with annual chill portions.}

\item{min_chill}{is a value of the lowest expected amount of chill for which the yield should be estimated (must be < expectedhighchill).}

\item{max_chill}{is a value of the highest expected  amount of chill for which the yield should be estimated (must be > expectedlowchill).}
}
\description{
Plot representing probabilities (shown along the y-axis) for the expected yield (shown along the x-axis). 
This is a broad slice through the density kernel from pasitR::chillkernel() function, which integrates to 1, the probability values are relative, not absolute measures.
}
\examples{
chill <- sample(x = 1:50, size = 20, replace = TRUE)
yield <- sample(x = 1000:5000, size = 20, replace = TRUE)
chillkernelslicerange(chill, yield, 10, 20)

}
\keyword{chill}
\keyword{density}
\keyword{kernel}
\keyword{yield}
