% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{Chillscatter}
\alias{Chillscatter}
\title{Create a scatter plot of chill portions and yield}
\usage{
Chillscatter(chill, yield)
}
\arguments{
\item{chill}{is a list of chill portions corresponding to another list with yields.}

\item{yield}{is a list of obsrved yields corresponding to another list with chill portions.}
}
\description{
Chill portions (x) and yield (y) scatter plot with associated and estimated densities  (using a loess smooth) given by the red dot (mean) and red ellipses (1 and 2 sigma from mean). The red line going across the plot shows the linear fit. Histograms are shown with smooth lines (loess smooth linear fits) density curves. The numeric value in the upper right gives the Spearman correlation coefficient between chill portions and yield.
}
\examples{

Chillscatter(chilldata$Chill_portions, chilldata$Yield)

}
\keyword{chill}
\keyword{chill,}
\keyword{portions,}
\keyword{yield}
