% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillviolin.R
\name{chillviolin}
\alias{chillviolin}
\title{Create a scatter plot of chill and yield}
\usage{
chillviolin(chill, yield)
}
\arguments{
\item{chill}{is a list of observed annual chill portions corresponding to another list with annual yields.}

\item{yield}{is a list of observed annual yields corresponding to another list with annual chill portions.}
}
\description{
Violin plots of chill portions (x) and yield (y) with six different intervals of chill portions.
Determine different possible chill portion intervals by calculating the optimal interval width for chill portions using the IQR() function in the stats() package of R, after the Freedman-Diaconis rule (IQR = interquartile range).
}
\examples{
chill <- sample(x = 1:50, size = 20, replace = TRUE)
yield <- sample(x = 1000:5000, size = 20, replace = TRUE)
chillviolin(chill, yield)

}
\keyword{Freedman-Diaconis}
\keyword{IQR}
\keyword{chill}
\keyword{violin}
\keyword{yield}
