% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillscatter.R
\name{chillscatter}
\alias{chillscatter}
\title{Create a scatter plot of chill and yield}
\usage{
chillscatter(chill, yield)
}
\arguments{
\item{chill}{is a vector of observed annual chill portions corresponding to another list with annual yields.}

\item{yield}{is a vector of observed annual yields corresponding to another list with annual chill portions.}
}
\description{
Chill (x) and yield (y) scatter plot with associated and estimated densities  (using a loess smooth) given by the red dot (mean) and red ellipses (1 and 2 sigma from mean). The red line going across the plot shows the linear fit. Histograms are shown with smooth lines (loess smooth linear fits) density curves. The numeric value in the upper right gives the Spearman correlation coefficient between chill portions and yield.
}
\examples{
chill <- sample(x = 1:50, size = 20, replace = TRUE)
yield <- sample(x = 1000:5000, size = 20, replace = TRUE)
chillscatter(chill, yield)

}
\keyword{chill}
\keyword{chill,}
\keyword{portions,}
\keyword{yield}
